// src/utils/types/TechnopalApiObject.ts
export type TechnopalApiObject = {
    program_description: string;
    user_group: string;
    address: string;
    is_active: boolean;
    gender: string;
    last_name: string;
    section: string;
    middle_name: string;
    login: boolean;
    year_level: number;
    email_address: string;
    user_code: string;
    program_code: string;
    err_msg: string;
    is_valid: boolean;
    cp_number: string;
    rfid: string;
    first_name: string;
}