// src/utils/types/User.ts
type User = {
	id_number: string;
	firstname: string;
	lastname: string;
	middlename: string;
	email: string;
	cp_number: string;
	course: string;
	year_level: number;
	section: string;
	program_description: string;
	is_active: number;
	user_group: string;
	password: string;
	year_active: number;
	created_at: string;
};

export { User };